cleaned up some css hopefully, made it so user can't  do unaothrized actions because I know Javier is gonna try to hack the this app

+++ frontend/src/components/Albums/album.js
@@ -12,13 +12,22 @@ function Album() {
     const dispatch = useDispatch();
     const history = useHistory();
 
+    const userId = photoArr.filter(photo => sessionUser?.id === photo?.id);
+
     useEffect(() => {
         dispatch(getAlbum(id));
     }, [dispatch, id]);
 
     const handleDeletAlbum = async () => {
-        await dispatch(deleteAlbum(id))
-        history.push('/album')
+        if (userId !== null ) {
+            await dispatch(deleteAlbum(id))
+            history.push('/album')
+        }
+        else {
+            <h1>
+                Seriously, stop trying to hack the website
+            </h1>
+        }
     }
 
     function newDirect() {
@@ -30,7 +39,7 @@ function Album() {
     return (
         <div>
             <div className="one__album__div">
-                <button className="new__photo__button" onClick={newDirect}>Add Photo</button>
+                <button hidden={sessionUser.id !== Number(id) ? true : false}  className="new__photo__button" onClick={newDirect}>Add Photo</button>
                 {photoArr?.map(photo => {
                     if (photo?.albumId === Number(id)) {
                         return (
@@ -38,7 +47,7 @@ function Album() {
                                 <div>
                                     {/* <h2 key={photo?.name}>{photo?.name}</h2> */}
                                     <Link to={`/photo/${photo?.id}`}>
-                                        <img key={photo?.id} src={photo?.imgUrl} alt="meow" height="100" width="140"></img>
+                                        <img key={photo?.albumId} src={photo?.imgUrl} alt="meow" height="200" width="230"></img>
                                     </Link>
                                 </div>
                             </div>
